Create a delivery_services.go file and move the rest urisigning handler functions to this file.  Also, since riak is optional, add a flag, RiakEnabled, to the config to inform api endpoints that use riak wether or not riak is used.
